+++ CMakeLists.txt
@@ -20,9 +20,16 @@ catkin_package(CATKIN_DEPENDS
 
 include_directories(
   ${catkin_INCLUDE_DIRS}
+  ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
 
-add_executable(pacmod_game_control_node src/pacmod_game_control_node.cpp)
+add_executable(pacmod_game_control_node 
+  src/pacmod_game_control_node.cpp
+  src/startup_checks.cpp
+  src/publish_control_board_rev2.cpp
+  src/publish_control_board_rev3.cpp
+)
+
 add_dependencies(pacmod_game_control_node
   ${${PROJECT_NAME}_EXPORTED_TARGETS}
   ${catkin_EXPORTED_TARGETS}
