+++ include/startup_checks.h
@@ -1,0 +1,20 @@
+/*
+* Unpublished Copyright (c) 2009-2018 AutonomouStuff, LLC, All Rights Reserved.
+*
+* This file is part of the PACMod ROS 1.0 driver which is released under the MIT license.
+* See file LICENSE included with this software or go to https://opensource.org/licenses/MIT for full license details.
+*/
+/*
+ * Series of startup checks for the pacmod game control node.
+ */
+
+
+#ifndef STARTUP_CHECKS_H
+#define STARTUP_CHECKS_H
+
+#include "globals.h"
+
+
+bool run_startup_checks_ok(ros::NodeHandle * nodeH);
+
+#endif
